Renamed folders to tamtt
