+++ tamtt/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TamttConfig(AppConfig):
+    name = 'TAMTT'
