+++ tamtt/templates/tamtt/calendar.html
@@ -2,7 +2,7 @@
       -->
 
 <!-- <CalendarSnippet> -->
-{% extends "tutorial/layout.html" %}
+{% extends "tamtt/layout.html" %}
 {% block content %}
 {% load custom_tags %}
 <h1>Calendar</h1>
