+++ tamtt/templates/tamtt/home.html
@@ -1,8 +1,5 @@
-<!-- Copyright (c) Microsoft Corporation.
-     Licensed under the MIT License. -->
-
 <!-- <HomeSnippet> -->
-{% extends "tutorial/layout.html" %}
+{% extends "tamtt/layout.html" %}
 {% block content %}
 <div class="jumbotron">
   <h1>TAMTT</h1>
