+++ tamtt/templates/tamtt/layout.html
@@ -12,7 +12,7 @@
       crossorigin="anonymous">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.13.0/css/all.css">
     {% load static %}
-    <link rel="stylesheet" href="{% static "tutorial/app.css" %}">
+    <link rel="stylesheet" href="{% static "tamtt/app.css" %}">
   </head>
 
   <body>
