+++ tamtt/templates/tamtt/presence.html
@@ -1,6 +1,6 @@
 
 <!-- <PresenceSnippet> -->
-{% extends "tutorial/layout.html" %}
+{% extends "tamtt/layout.html" %}
 {% block content %}
 <h1>Presence</h1>
 <table class="table">
