+++ tamtt_app/settings.py
@@ -38,7 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'tutorial',
+    'tamtt',
 ]
 # </InstalledAppsSnippet>
